Get All Install Apps with Description
This script generates a CSV file with a list of all the installed applications on a Windows machine, along with a description of what each application does.

To use the script, simply copy and paste the code into a PowerShell window and run it. The CSV file will be created in the root of the C: drive and will be named "installed_apps.csv".

The script uses the Get-WmiObject cmdlet to get a list of all installed apps on the machine and the Export-Csv cmdlet to export the list to a CSV file. The script also creates a new object for each app to store its name and description and adds the object to an array which is then exported to the CSV file.

This script can be useful for generating a list of installed apps for inventory purposes or for troubleshooting issues with specific applications.