Add files via upload
This script uses the Get-WmiObject cmdlet to retrieve a list of all installed drivers on the machine. It then iterates through the list of drivers, creating a new object for each driver to store its name, manufacturer, date, and version. The script then exports the list of drivers to a CSV file using the Export-Csv cmdlet. This allows you to easily see a list of all the drivers installed on the machine, as well as their relevant information.