Create PowerShell Script for Exporting GPOs with Keyword Filter
This PowerShell script allows administrators to filter GPOs based on a keyword entered by the user, export the filtered GPOs' properties to a CSV file and schedule it to run at regular intervals. It can save time and effort when managing GPOs in a Windows domain and can also be used for monitoring and tracking GPOs over a period of time.